Create ToggleButton tsx file and combine the different styled components.
